checkout changes from slide_dataloader

+++ hi-ml-histopathology/src/histopathology/models/deepmil.py
@@ -226,7 +226,7 @@ class DeepMILModule(LightningModule):
 
         results = dict()
         for metric_object in self.get_metrics_dict(stage).values():
-            metric_object.update(predicted_probs, bag_labels.view(batch_size,))
+            metric_object.update(predicted_probs, bag_labels.view(batch_size,).int())
         results.update({ResultsKey.SLIDE_ID: batch[TilesDataset.SLIDE_ID_COLUMN],
                         ResultsKey.TILE_ID: batch[TilesDataset.TILE_ID_COLUMN],
                         ResultsKey.IMAGE_PATH: batch[TilesDataset.PATH_COLUMN],
